+++ src/main/java/site/radio/clova/dailyReport/DailyReportMessageParser.java
@@ -1,10 +1,15 @@
 package site.radio.clova.dailyReport;
 
+import java.util.List;
+import java.util.stream.Collectors;
 import site.radio.clova.dto.ClovaResponseDto;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import site.radio.letter.Letter;
 
-public class DailyReportExtractor {
+public class DailyReportMessageParser {
+
+    private static final String LETTER_SEPARATOR = "sharpie-sep";
 
     public static ClovaDailyAnalysisResult extract(ClovaResponseDto response) {
         ObjectMapper objectMapper = new ObjectMapper();
@@ -32,5 +37,25 @@ public class DailyReportExtractor {
             }
         }
     }
-}
 
+    public static String requestClovaAnalysis(List<Letter> letters) {
+        // 편지 내용 구분자 동적 생성
+        String msgSeparator = Long.toHexString(Double.doubleToLongBits(Math.random()));
+
+        return letters.stream()
+                .map(letter -> String.format("<%s:%s>\n%s\n</%s:%s>",
+                        LETTER_SEPARATOR, msgSeparator,
+                        reformatMsg(letter.getMessage()),
+                        LETTER_SEPARATOR, msgSeparator))
+                .collect(Collectors.joining("\n"));
+    }
+
+    private static String reformatMsg(String input) {
+        return input
+                .replaceAll("<", "&lt;")
+                .replaceAll(">", "&gt;")
+                .replaceAll("&", "&amp;")
+                .replaceAll("\"", "&quot;")
+                .replaceAll("'", "&apos;");
+    }
+}
